FpDebug: Fix overload compile error with fpc 3.2 / DestSize can not exceed the int64 range
git-svn-id: trunk@62019 -

+++ components/fpdebug/fpdmemorytools.pas
@@ -1299,7 +1299,7 @@ begin
 
 
         if SourceReadSize <= ConvData.DestSize then begin
-          move(ReadData^, ADest^, Min(SizeOf(TmpVal) ,ConvData.DestSize)); // Little Endian only
+          move(ReadData^, ADest^, Min(SizeOf(TmpVal), Int64(ConvData.DestSize))); // Little Endian only
           ReadData := ADest;
         end;
